Avance en clases: varias fn, cambios en main y lista

+++ alumno.py
@@ -9,13 +9,13 @@ class Alumno():
         ponderado = self.promedio * self.cantidad_notas
         ponderado += nota
         self.cantidad_notas += 1
-        self.promedio = ponderado / self.cantidad_notas
+        self.promedio = int(ponderado / self.cantidad_notas)
 
     def quitar_nota(self, nota):
         ponderado = self.promedio * self.cantidad_notas
         ponderado -= nota
         self.cantidad_notas -= 1
-        self.promedio = ponderado / self.cantidad_notas
+        self.promedio = int(ponderado / self.cantidad_notas)
 
     def set_nombre(self, nombre):
         self.nombre =  nombre
